--- lib/utils/utils.dart
@@ -1,1 +1,0 @@
-class Utils {}
